Update README.md
This fast food ordering app is built using java. The menu is from meals, sandwich, dessert, drinks and shakes. A customer chooses the type of payment, that is after choosing the food. They write the amount of money they want to pay with, a tax will be included then after the system will calculate all that and give out change.